DAAI Project Folder Structure/requirements.txt
open-clip-torch
tqdm
scipy